Remove MaterialIcons-Regular.woff2 from repository
material-design-icons requires that css and webfont are in the same directory.
In rails, css is compiled to `public/assets` by sprockets so webfont
required to put in `public/assets` directory **without digest**.

```css
@font-face {
  font-family: 'Material Icons';
  font-style: normal;
  font-weight: 400;
  src: url(MaterialIcons-Regular.eot); /* For IE6-8 */
  src: local('Material Icons'),
       local('MaterialIcons-Regular'),
       url(MaterialIcons-Regular.woff2) format('woff2'),
       url(MaterialIcons-Regular.woff) format('woff'),
       url(MaterialIcons-Regular.ttf) format('truetype');
}
```

BTW, `public/assets` is usually set to `linked_dir` in capistrano.
Do not include MaterialIcons-Regular.woff2 in repository and put it in server's `shared/public/assets`.